Chore: components, store 폴더 구조 변경
