+++ src/components/common/CreateInvitation/InvitationTitleInput.tsx
@@ -19,7 +19,7 @@ const InvitationTitleInput: React.FC<InvitationTitleInputProps> = ({
         placeholder="제목을 입력해주세요"
         required
         onChange={onChange}
-        className="px-4 py-2  text-xs text-center text-slate-500  bg-slate-300 bg-opacity-30 rounded-xl border-none placeholder-slate-500 placeholder-opacity-50 focus:outline-none focus-visible:ring-1 focus:ring-slate-300"
+        className="px-4 py-2 text-xs text-center text-slate-500  bg-sky-200 bg-opacity-30 rounded-xl shadow-sm border-none placeholder-slate-500 placeholder-opacity-60 focus:outline-none focus-visible:ring-1 focus:ring-slate-400"
       ></input>
     </div>
   );
