+++ src/components/form/NameInput.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import useBrideGroomStore from './useBrideGroomStore';
+import useBrideGroomStore from '../../store/useBrideGroomStore';
 
 const NameInput = () => {
   const brideGroom = useBrideGroomStore((state) => state.brideGroom);
