+++ src/components/form/WeddingDateInput/TimePicker.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDatePicker, { registerLocale } from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
 import { ko } from 'date-fns/locale/ko';
-import { useWeddingStore } from '../../store/store';
+import { useWeddingStore } from '../../../store/useWeddingStore';
 
 const TimePicker: React.FC = () => {
   const { weddingTime, setWeddingTime } = useWeddingStore();
