+++ src/components/layout/PageLayout.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from './Header';
+import Header from '../common/Header/Header';
 
 type PageLayoutProps = {
   title?: string;
