+++ src/components/layout/StepNavigation.tsx
