+++ src/store/useWeddingStore.ts
@@ -14,14 +14,3 @@ export const useWeddingStore = create<WeddingStore>((set) => ({
   weddingTime: '12:00',
   setWeddingTime: (time) => set({ weddingTime: time }),
 }));
-
-// 청첩장 정보 - 제목
-type InvitationStore = {
-  title: string;
-  setTitle: (newTitle: string) => void;
-};
-
-export const useInvitationStore = create<InvitationStore>((set) => ({
-  title: '',
-  setTitle: (newTitle) => set({ title: newTitle }),
-}));
